programmer/requirements.txt
pyserial==3.5
intelhex==2.3.0
